+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to
   - Add support of slurm-quota (#52).
 - cli: Add `deploy --update-os-image` option to force download of base OS image
   when already present on host.
+- load: Add random yet realistic job names.
 - lib: Add `deploy --update-os-image` option in bash-completion.
 - docs: Mention `deploy --update-os-image` option in manpage.